Analyze Multiple Images

Detect the water-sensitive paper from multiple images and do all the analyses in a run.

analyzeImages(path, paper_dim = c(76, 26), distance = 0.7, display.it = FALSE)

Arguments

path
A character giving the path to the folder containing the images (only!) of the papers.

paper_dim
A numeric vector of length 2 indicating the length and width (in mm) of the water-sensitive paper.

distance
A numeric value indicating the distance from the nozzle to the target, in order to estimate the spray drift (%) using the 'German model' for field crops.

display.it
(Logical) Should the image with the bounding box of the paper be displayed? Default is TRUE.

Value

A list of objects of class hydropaper. See more in analyzePaper().

Examples

path <- system.file('images', package = 'hydropaper')
list.files(path)
#> [1] "field1.JPG" "field2.JPG"
analyzeImages(path)
#> $field1.JPG
#> 
#>               Water-Sensitive Paper Analysis
#>  
#>                             N drops: 1394 
#>                  Spray coverage (%): 18.1 
#>                 Density (drops/cm2): 71 
#>          Vol. applied (microliters): 213.44 
#>                    L/ha (predicted): 1080.15 
#>        Mean diameter (micronmeters): 462 
#>                   NMD (micrometers): 357 
#>                   VMD (micrometers): 1042 
#>                   D.1 (micrometers): 608 
#>                   D.9 (micrometers): 1456 
#>                  Relative amplitude: 0.81 
#>          Largest drop (micrometers): 1553 
#>         Smallest drop (micrometers): 113 
#> Diam. class % (<105, 105-340, >340): 0 49 51 
#>                         CV area (%): 135.2 
#>                        CV diam. (%): 72.8 
#>                       Good to spray: 
#>                           Drift (%): 96.1 
#> 
#> 
#> $field2.JPG
#> 
#>               Water-Sensitive Paper Analysis
#>  
#>                             N drops: 43 
#>                  Spray coverage (%): 0.5 
#>                 Density (drops/cm2): 2 
#>          Vol. applied (microliters): 6.18 
#>                    L/ha (predicted): 31.29 
#>        Mean diameter (micronmeters): 390 
#>                   NMD (micrometers): 271 
#>                   VMD (micrometers): 1545 
#>                   D.1 (micrometers): 606 
#>                   D.9 (micrometers): 1665 
#>                  Relative amplitude: 0.69 
#>          Largest drop (micrometers): 1665 
#>         Smallest drop (micrometers): 136 
#> Diam. class % (<105, 105-340, >340): 0 58.1 41.9 
#>                         CV area (%): 207.2 
#>                        CV diam. (%): 87 
#>                       Good to spray: 
#>                           Drift (%): 96.1 
#> 
#>